refactor upload page by merging two components into one components, enhance readable
